added scalable routing
