+++ app.js
@@ -1,0 +1,36 @@
+const express = require('express')
+const bodyParser = require('body-parser')
+const bcrypt = require('bcrypt')
+const app = express();
+const mountRoutes = require('./routes');
+
+app.use(bodyParser.json())
+app.use(bodyParser.urlencoded({ extended: true }))
+
+/* cross origin */
+app.use(function(req, res, next) {
+  res.header('Access-Control-Allow-Origin', '*');
+  res.header('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE,OPTIONS');
+  res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With');
+  next();
+});
+
+mountRoutes(app);
+
+app.use(function(req, res, next) {
+	const err = new Error('Not Found');
+	err.status = 404;
+	next(err);
+});
+
+app.use(function(err, req, res, next) {
+	// set locals, only providing error in development
+	res.locals.message = err.message;
+	res.locals.error = req.app.get('env') === 'development' ? err : {};
+
+	// render the error page
+	res.status(err.status || 500);
+	res.render('error');
+});
+
+app.listen(3000);
