+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
+    "start": "nodemon app.js",
     "build": "node models/database.js"
   },
   "author": "",
@@ -14,6 +14,7 @@
     "bcrypt": "^2.0.1",
     "body-parser": "^1.18.2",
     "express": "^4.16.3",
+    "express-promise-router": "^3.0.2",
     "nodemon": "^1.17.3",
     "nvm": "0.0.4",
     "pg": "^7.4.1"
