+++ routes/index.js
@@ -1,0 +1,8 @@
+const apps = require('./apps');
+const admin = require('./admin');
+
+module.exports = (app) => {
+	// app.use('/', (req, res) => res.render('index', {} ))
+	app.use('/apps', apps);
+	app.use('/admin', admin);
+};
